Pass pre-expanded media tokens to MLLM.
