+++ tests/unit/models/generation/test_megatron_generation.py
@@ -28,6 +28,7 @@ from nemo_rl.models.generation.megatron import MegatronGeneration, megatron_gene
 from nemo_rl.models.generation.megatron.config import (
     dedicated_inference_megatron_cfg,
 )
+from nemo_rl.models.generation.megatron.megatron_worker import MegatronGenerationMixin
 from nemo_rl.models.policy import PolicyConfig
 from nemo_rl.models.policy.lm_policy import Policy
 from nemo_rl.weight_sync.megatron_weight_synchronizer import (
@@ -36,6 +37,67 @@ from nemo_rl.weight_sync.megatron_weight_synchronizer import (
 from tests.unit.test_utils import SimpleLossFn
 
 model_name = "Qwen/Qwen3-0.6B"
+
+
+def test_direct_megatron_media_request_preserves_preexpanded_prompt():
+    worker = object.__new__(MegatronGenerationMixin)
+    worker.cfg = {"generation": {"mcore_generation_config": {}}}
+
+    def sample_vision_tensors(data, index):
+        return torch.ones(1, 2, 4), torch.tensor([[2, 2]]), None
+
+    def get_wrapper_cls():
+        return object
+
+    def wrapper_supports_modality(wrapper_cls, modality):
+        return modality == "image"
+
+    worker._sample_vision_tensors = sample_vision_tensors
+    worker._get_megatron_inference_wrapper_cls = get_wrapper_cls
+    worker._wrapper_supports_modality = wrapper_supports_modality
+    data = {
+        "input_ids": torch.tensor([[10, 99, 99, 20, 0]]),
+        "input_lengths": torch.tensor([4]),
+    }
+
+    prompt, multi_modal_data = worker._build_prompt_and_multimodal_data(data, 0)
+
+    assert prompt == [10, 99, 99, 20]
+    assert multi_modal_data["media_tokens_preexpanded"] is True
+    assert "image" in multi_modal_data
+
+
+def test_direct_megatron_video_request_marks_preexpanded_prompt():
+    worker = object.__new__(MegatronGenerationMixin)
+    worker.cfg = {"generation": {"mcore_generation_config": {}}}
+
+    def sample_vision_tensors(data, index):
+        return (
+            torch.ones(1, 4, 4),
+            torch.tensor([[2, 2], [2, 2], [2, 2], [2, 2]]),
+            torch.tensor([4]),
+        )
+
+    def get_wrapper_cls():
+        return object
+
+    def wrapper_supports_modality(wrapper_cls, modality):
+        return modality == "video"
+
+    worker._sample_vision_tensors = sample_vision_tensors
+    worker._get_megatron_inference_wrapper_cls = get_wrapper_cls
+    worker._wrapper_supports_modality = wrapper_supports_modality
+    data = {
+        "input_ids": torch.tensor([[10, 99, 99, 20]]),
+        "input_lengths": torch.tensor([4]),
+    }
+
+    prompt, multi_modal_data = worker._build_prompt_and_multimodal_data(data, 0)
+
+    assert prompt == [10, 99, 99, 20]
+    assert multi_modal_data["media_tokens_preexpanded"] is True
+    assert "video" in multi_modal_data
+
 
 basic_megatron_test_config: PolicyConfig = {
     "model_name": model_name,
